requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
channels==4.0.0
click==8.1.7
constantly==23.10.4
coverage==7.5.4
decorator==5.1.1
Django==5.0.6
django-cors-headers==4.4.0
django-extensions==3.2.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-taggit==5.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
executing==2.0.1
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.1
iniconfig==2.0.0
ipython==8.26.0
jedi==0.19.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
numpy==1.26.4
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg==3.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest_httpserver==1.0.10
python-slugify==8.0.4
setuptools==72.0.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.30.3
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
zope.interface==6.4.post2
